Fix simplify imports for PinterestLayout
